// backend/src/category/enums/category.enum.ts
export enum CATEGORY_CODE {
  CITY = 'CITY',
  COMMUNITY_EVENTS = 'COMMUNITY_EVENTS',
  CRIME_AND_SAFETY = 'CRIME_AND_SAFETY',
  CULTURE = 'CULTURE',
  DISCOUNTS_AND_BENEFITS = 'DISCOUNTS_AND_BENEFITS',
  EMERGENCIES = 'EMERGENCIES',
  FOR_SENIORS = 'FOR_SENIORS',
  HEALTH = 'HEALTH',
  KIDS_AND_FAMILY = 'KIDS_AND_FAMILY',
}